+++ lua/obsidian-para-flow/ui.lua
@@ -1,5 +1,6 @@
 local M = {}
 local input
+local select
 
 local ReviewView = {}
 ReviewView.__index = ReviewView
@@ -138,6 +139,11 @@ function M.input(options, callback)
   handler(options, callback)
 end
 
+function M.select(items, options, callback)
+  local handler = select or vim.ui.select
+  handler(items, options, callback)
+end
+
 function M.notify_error(message)
   vim.notify("obsidian-para-flow: " .. message, vim.log.levels.ERROR)
 end
@@ -187,8 +193,13 @@ function M._set_input(value)
   input = value
 end
 
+function M._set_select(value)
+  select = value
+end
+
 function M._reset()
   input = nil
+  select = nil
 end
 
 return M